Adds supeclass for marker types. Checkerboard type now also supports generating robot configy by range values.
