+++ robotino_calibration/ros/include/robotino_calibration/camera_base_calibration_checkerboard.h
@@ -51,59 +51,27 @@
 #ifndef __CAMERA_BASE_CALIBRATION_CHECKERBOARD_H__
 #define __CAMERA_BASE_CALIBRATION_CHECKERBOARD_H__
 
-// ROS
-#include <ros/ros.h>
 
-#include <tf/transform_listener.h>
-#include <sensor_msgs/Image.h>
-#include <geometry_msgs/Twist.h>
-
-// image transport
-#include <image_transport/image_transport.h>
-#include <image_transport/subscriber_filter.h>
-
-// OpenCV
-#include <opencv/cv.h>
-
-// PCL
-#include <pcl/point_cloud.h>
-#include <pcl/point_types.h>
-
-// opencv
-#include <opencv/cv.h>
-#include <opencv/highgui.h>
-#include <cv_bridge/cv_bridge.h>
-#include <sensor_msgs/image_encodings.h>
-
-// Boost
-#include <boost/thread/mutex.hpp>
-
-#include <robotino_calibration/calibration_utilities.h>
+#include <robotino_calibration/camera_base_calibration_marker.h>
 
 
-class CameraBaseCalibrationCheckerboard
+class CameraBaseCalibrationCheckerboard : public CameraBaseCalibrationMarker
 {
 public:
 
 	CameraBaseCalibrationCheckerboard(ros::NodeHandle nh);
-
 	~CameraBaseCalibrationCheckerboard();
 
 	// starts the calibration between camera and base including data acquisition
 	bool calibrateCameraToBase(const bool load_images);
 
-	void setCalibrationStatus(bool calibrated)
-	{
-		calibrated_ = calibrated;
-	}
-
 	// load/save calibration data from/to file
 	bool saveCalibration();
 	bool loadCalibration();
-
 	void getCalibration(cv::Mat& K, cv::Mat& distortion, cv::Mat& T_base_to_torso_lower, cv::Mat& T_torso_upper_to_camera);
 
 	void undistort(const cv::Mat& image, cv::Mat& image_undistorted);
+
 
 protected:
 
@@ -123,32 +91,13 @@ protected:
 	int acquireCalibrationImage(int& image_width, int& image_height, std::vector<cv::Point2f>& points_2d_per_image,
 			const cv::Size pattern_size, const bool load_images, int& image_counter);
 
-	// moves the robot to a desired location and adjusts the torso joints
-	bool moveRobot(const RobotConfiguration& robot_configuration);
-
 	// generates the 3d coordinates of the checkerboard in local checkerboard frame coordinates
 	void computeCheckerboard3dPoints(std::vector< std::vector<cv::Point3f> >& pattern_points, const cv::Size pattern_size, const double chessboard_cell_size, const int number_images);
 
 	// intrinsic camera calibration (+ distortion coefficients)
 	void intrinsicCalibration(const std::vector< std::vector<cv::Point3f> >& pattern_points, const std::vector< std::vector<cv::Point2f> >& camera_points_2d_per_image, const cv::Size& image_size, std::vector<cv::Mat>& rvecs_jai, std::vector<cv::Mat>& tvecs_jai);
 
-	void extrinsicCalibrationBaseToTorsoLower(std::vector< std::vector<cv::Point3f> >& pattern_points_3d,
-			std::vector<cv::Mat>& T_base_to_checkerboard_vector, std::vector<cv::Mat>& T_torso_lower_to_torso_upper_vector,
-			std::vector<cv::Mat>& T_camera_to_checkerboard_vector);
 
-	void extrinsicCalibrationTorsoUpperToCamera(std::vector< std::vector<cv::Point3f> >& pattern_points_3d,
-			std::vector<cv::Mat>& T_base_to_checkerboard_vector, std::vector<cv::Mat>& T_torso_lower_to_torso_upper_vector,
-			std::vector<cv::Mat>& T_camera_to_checkerboard_vector);
-
-	// computes the rigid transform between two sets of corresponding 3d points measured in different coordinate systems
-	// the resulting 4x4 transformation matrix converts point coordinates from the target system into the source coordinate system
-	cv::Mat computeExtrinsicTransform(const std::vector<cv::Point3d>& points_3d_source, const std::vector<cv::Point3d>& points_3d_target);
-
-	// computes the transform from target_frame to source_frame (i.e. transform arrow is pointing from target_frame to source_frame)
-	bool getTransform(const std::string& target_frame, const std::string& source_frame, cv::Mat& T);
-
-
-	ros::NodeHandle node_handle_;
 	image_transport::ImageTransport* it_;
 	image_transport::SubscriberFilter color_image_sub_; ///< Color camera image input topic
 	boost::mutex camera_data_mutex_;	// secures read and write operations on camera data
@@ -156,37 +105,11 @@ protected:
 	ros::Time latest_image_time_;	// stores time stamp of latest image
 	bool capture_image_;
 
-	ros::Publisher base_controller_;
-	ros::Publisher tilt_controller_;
-	ros::Publisher pan_controller_;
-
-	tf::TransformListener transform_listener_;
-	std::string torso_lower_frame_;
-	std::string torso_upper_frame_;
-	std::string camera_frame_;
-	std::string camera_optical_frame_;
-	std::string base_frame_;
-	std::string checkerboard_frame_;
-
 	cv::Mat K_;			// intrinsic matrix for camera
 	cv::Mat distortion_;	// distortion parameters for camera
 
-	cv::Mat T_base_to_torso_lower_;		// transformation to estimate from base to torso_lower
-	cv::Mat T_torso_upper_to_camera_;		// transformation to estimate from torso_upper to camera
-
-	bool calibrated_;	// only true if cameras were calibrated or a calibration was loaded before
-
-	// parameters
-	std::string camera_calibration_path_;	// path to data
-	std::string tilt_controller_command_;
-	std::string pan_controller_command_;
-
 	double chessboard_cell_size_;	// cell side length in [m]
 	cv::Size chessboard_pattern_size_;		// number of checkerboard corners in x and y direction
-
-	int optimization_iterations_;	// number of iterations for optimization
-
-	std::vector<RobotConfiguration> robot_configurations_;	// list of robot configurations for observing the checkerboard
 };
 
 #endif // __CAMERA_BASE_CALIBRATION_CHECKERBOARD_H__
