+++ robotino_calibration/ros/launch/camera_base_calibration_pitag_params.yaml
@@ -7,7 +7,7 @@ T_base_to_torso_lower_initial: [0.25, 0, 0.5, 0.0, 0.0, 0.0]
 T_torso_upper_to_camera_initial: [0.0, 0.065, 0.0, 0.0, 0.0, -1.57]
 
 
-### checkerboard observation positions for capturing calibration images
+### Observation positions for capturing calibration images of the pitags
 # if this flag is true, then the robot configurations will be sampled on a equally spaced grid from the given ranges, if false, the user-provided configurations in robot_configurations will be used
 # bool
 use_range: true
@@ -21,7 +21,7 @@ phi_range: [0.0, 1.0, 0.0]         # in [rad], fixed phi-angle
 pan_range: [-0.52, 0.02, 0.52]     # in [rad]
 tilt_range: [-0.44, 0.02, 0.44]    # in [rad]
 
-# list of robot configurations for observing the checkerboard measured relative to the checkerboard_reference_nav coordinate system that is similarly aligned as the robot's base_link facing the checkerboard
+# list of robot configurations for observing the pitags measured relative to the landmark_reference_nav coordinate system that is similarly aligned as the robot's base_link facing the marker
 # includes 5 parameters per entry: robot pose: x, y, phi and torso: pan, tilt
 robot_configurations: [-1.5, -0.17, 0, 0.15, 0.25,
                        -1.5, -0.17, 0, 0.0, 0.3,
@@ -75,7 +75,7 @@ camera_optical_frame: "kinect_rgb_optical_frame"
 # the robot base frame [the transformation between laser scanner and base should be accomplished before, the transform from base_frame to torso_lower_framewill be calibrated by this program]
 base_frame: "base_link"
 
-# do not modify, this is the coordinate system fixed in the upper left checkerboard calibration corner (where 4 squares meet), it will be published to tf via the checkerboard_localisation program
+# base name of all frames generated by pitags, each pitag number will be attached to this string.
 marker_frame_base_name: "marker"
